+++ src/nexuralnet/dnn/utility/helper.h
@@ -30,7 +30,7 @@ namespace nexural {
 	namespace helper {
 		std::vector<std::string> TokenizeString(const std::string& str, const std::string& delimiters);
 		void SplitData(const Tensor& data, Tensor &firstFold, Tensor &secondFold, const float_n firstFoldPercentage = 90);
-		void BestClassClassification(const Tensor& tensor, size_t& bestClass);
+		void BestClassClassification(const Tensor& tensor, long& bestClass);
 		std::string NetworkResultTypeToString(const NetworkResultType netResultType);
 		float_n clip(const float_n& n, const float_n& lower, const float_n& upper);
 	}
